pyproject.toml
[project]
name = "sensorium"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "paho-mqtt>=2.1.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
]
